requirements.txt
pyVmomi
python-dotenv